Complete CheckWin and 95% complete of integration test
For CheckWin, merge and update with newest score algorithm with proper
exception.

For integration test, need to double check on exception list if
anythings else need to change/edit.

Updated Unit test case of testcheckwin to get 100% coverage (expcetion
checkscore, as it is integration with other class, did it in integration
test with 100% of checkscore)